requirements.txt
argparse==1.4.0
numpy==1.21.0
pathlib==1.0.1
h5py==3.6.0
pyYAML==6.0
sklearn==0.0
einops==0.5.0
seaborn==0.12.2
timm==0.8.0.dev0
asteroid==0.6.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cu113
torchvision==0.12.0+cu113
torchaudio==0.11.0+cu113